Put together the first stable version of the game
The first steps were getting the right factories up and running in a
TDD environment, as well as laying down the foundations of the game
flow.

Afterwards, it was a matter of making the user interface talk
with those factories, and split up the preparation stage,
i.e. ship placement time, from the attack stage, as well as adhere to
the turn system.

The icing on the cake was the Drag an Drop feature so as to enhance
the ship placement experience.